[FBW] Added variables to support YourControls
The following L-variables have been introduced:
- `A32NX_SIDESTICK_LEFT_POSITION_<X|Y>`
- `A32NX_SIDESTICK_RIGHT_POSITION_<X|Y>`
- `A32NX_RUDDER_POSITION_OVERRIDE_ON`
- `A32NX_RUDDER_POSITION`
- `A32NX_THROTTLE_POSITION_OVERRIDE_ON`
- `A32NX_THROTTLE_POSITION_<1|2>`

L-Variables `A32NX_SIDESTICK_LEFT_POSITION_<X|Y>` are set to the
position of the local input.

L-Variables `A32NX_SIDESTICK_RIGHT_POSITION_<X|Y>` are never set.

L-Variables `A32NX_SIDESTICK_POSITION_<X|Y>` are the sum of both
sidesticks and are being used as input for the fly-by-wire system.

When `A32NX_RUDDER_POSITION_OVERRIDE_ON` is `0`, the variable
`A32NX_RUDDER_POSITION` is written with the local input. When it is
`1`, `A32NX_RUDDER_POSITION` is read and used as input for the
fly-by-wire system.

When `A32NX_THROTTLE_POSITION_OVERRIDE_ON` is `0`, the variable
`A32NX_THROTTLE_POSITION_<1|2>` are written with the local inputs.
When it is `1`, `A32NX_THROTTLE_POSITION_<1|2>` are read and used
as input.